Move CSVTranslationLookupService into Services dir
