+++ src/CSVTranslationLookup/CSV/CSVProcessor.cs
@@ -6,6 +6,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Text;
+using CSVTranslationLookup.Services;
 
 namespace CSVTranslationLookup.CSV
 {
